エフェクトの名前を「分岐とマップ合成」に変更
