+++ YukkuriMovieMaker.Plugin.Community.Shader/OutputMapComposite.hlsl
