+++ YukkuriMovieMaker.Plugin.Community.Shader/YukkuriMovieMaker.Plugin.Community.Shader.vcxproj.filters
@@ -108,7 +108,7 @@
 	<FxCompile Include="HeatHaze.hlsl">
       <Filter>ソース ファイル</Filter>
     </FxCompile>
-	<FxCompile Include="DynamicLerp.hlsl">
+	<FxCompile Include="OutputMapComposite.hlsl">
       <Filter>ソース ファイル</Filter>
     </FxCompile>
   </ItemGroup>
