+++ YukkuriMovieMaker.Plugin.Community/Effect/Video/OutputMapComposite/Texts.cs
@@ -1,6 +1,6 @@
 using YukkuriMovieMaker.Generator;
 
-namespace YukkuriMovieMaker.Plugin.Community.Effect.Video.DynamicLerp
+namespace YukkuriMovieMaker.Plugin.Community.Effect.Video.OutputMapComposite
 {
     [AutoGenLocalizer]
     partial class Texts
