+++ YukkuriMovieMaker.Plugin.Community/Effect/Video/OutputMapComposite/Texts.csv
@@ -1,0 +1,12 @@
+Key,Parameter,ja-jp,en-us,zh-cn,zh-tw,ko-kr,es-es,ar-sa,id-id
+OutputMapCompositeEffectName,,分岐とマップ合成,Map-Composite with Branch,与分支贴图合成,與分支貼圖合成,분기와 맵 합성,Componer con ramificación por mapa,تركيب مع التفريع بالخريطة,Gabungkan dengan Cabang dan Peta
+OutputMapCompositeTargetIndexName,,合成相手,Target Index,目标索引,目標索引,대상 인덱스,Índice objetivo,الفهرس المستهدف,Indeks Target
+OutputMapCompositeTargetIndexDesc,,補間先となる映像の分岐インデックスを指定します,Branch index of the image to lerp toward,指定插值目标映像的分支索引,指定插值目標影像的分支索引,보간 대상 영상의 분기 인덱스를 지정합니다,Índice de rama de la imagen hacia la que interpolar,فهرس الفرع للصورة المراد الإرجاع إليها,Indeks cabang dari gambar tujuan lerp
+OutputMapCompositeMapIndexName,,マップ映像,Map Index,贴图索引,貼圖索引,맵 인덱스,Índice del mapa,فهرس الخريطة,Indeks Peta
+OutputMapCompositeMapIndexDesc,,ブレンド重みを決定するマップ映像の分岐インデックスを指定します,Branch index of the image used to drive blend weight,指定用于确定混合权重的贴图映像分支索引,指定用於決定混合權重的貼圖影像分支索引,블렌드 가중치를 결정하는 맵 영상의 분기 인덱스를 지정합니다,Índice de rama de la imagen usada para determinar el peso de mezcla,فهرس الفرع للصورة المستخدمة لتحديد وزن المزج,Indeks cabang gambar yang menentukan bobot campuran
+OutputMapCompositeMapTypeName,,マップの種類,Map Type,贴图类型,貼圖類型,맵 종류,Tipo de mapa,نوع الخريطة,Jenis Peta
+OutputMapCompositeMapTypeDesc,,マップから読み取る値の種類を選択します,Select what value to read from the map,选择从贴图中读取的值类型,選擇從貼圖中讀取的值類型,맵에서 읽을 값의 종류를 선택합니다,Seleccione qué valor leer del mapa,حدد قيمة القراءة من الخريطة,Pilih jenis nilai yang dibaca dari peta
+OutputMapCompositeMapTypeLuminance,,輝度,Luminance,亮度,亮度,휘도,Luminancia,السطوع,Luminans
+OutputMapCompositeMapTypeLuminanceDesc,,マップ映像の輝度をブレンド重みとして使用します,Uses the luminance of the map image as blend weight,使用贴图亮度作为混合权重,使用貼圖亮度作為混合權重,맵 영상의 휘도를 블렌드 가중치로 사용합니다,Usa la luminancia del mapa como peso de mezcla,يستخدم سطوع صورة الخريطة كوزن للمزج,Menggunakan luminans gambar peta sebagai bobot campuran
+OutputMapCompositeMapTypeAlpha,,アルファ,Alpha,透明度,透明度,알파,Alfa,ألفا,Alfa
+OutputMapCompositeMapTypeAlphaDesc,,マップ映像のアルファ値をブレンド重みとして使用します,Uses the alpha of the map image as blend weight,使用贴图透明度作为混合权重,使用貼圖透明度作為混合權重,맵 영상의 알파 값을 블렌드 가중치로 사용합니다,Usa el alfa del mapa como peso de mezcla,يستخدم قيمة ألفا لصورة الخريطة كوزن للمزج,Menggunakan alpha gambar peta sebagai bobot campuran
